requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
google_api_python_client==2.129.0
jsonschema==4.22.0
langchain_anthropic==0.1.15
langchain_cohere==0.1.8
langchain_community==0.2.5
langchain_core==0.2.9
langchain_fireworks==0.1.3
langchain_mistralai==0.1.8
langchain_openai==0.1.8
mammoth==1.7.1
markdownify==0.12.1
pandas==2.2.2
plotly==5.22.0
pydantic==2.7.4
pymongo==4.7.2
pymongo[encryption]==4.7.2
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.4.2
st_combobox==1.1.1
streamlit==1.34.0
